clear code.
