+++ client/entities.go
@@ -151,32 +151,6 @@ type (
 		Message string
 	}
 
-	GroupSystemMessages struct {
-		InvitedRequests []*GroupInvitedRequest
-		JoinRequests    []*UserJoinGroupRequest
-	}
-
-	GroupInvitedRequest struct {
-		RequestId   int64
-		InvitorUin  int64
-		InvitorNick string
-		GroupCode   int64
-		GroupName   string
-
-		client *QQClient
-	}
-
-	UserJoinGroupRequest struct {
-		RequestId     int64
-		Message       string
-		RequesterUin  int64
-		RequesterNick string
-		GroupCode     int64
-		GroupName     string
-
-		client *QQClient
-	}
-
 	NewFriendRequest struct {
 		RequestId     int64
 		Message       string
